+++ CHANGELOG.md
@@ -13,6 +13,7 @@ _Not yet on NuGet..._
 * Coordinates: Added `Position` and `Coordinates` properties (#4185) @blouflashdb
 * Signal: Added `AlwaysUseLowDensityMode` for improved anti-aliased rendering in static plots (#4153)
 * Plot: Improved default `ToString()` implementation for the object returned when saving image files (#4154)
+* Plot: Added `Save()` overload which detects file format from filename extension (#4164)
 * Coordinates: Added `Coordinates.Zip()` for creating a `Coordinates[]` from distinct `xs[]` and `ys[]` arrays.
 * Lollipop: New plot type that displays values with a stem and a marker similar to a bar graph but with less visual clutter (#4193, #4183) @CoderPM2011